changing to JPA
